requirements.txt
onnxruntime==1.20.1
aiogram==3.15.0
g4f==0.3.9.5
python-dotenv==1.0.1
